+++ test_script.py
@@ -17,15 +17,13 @@ For an in-depth tutorial please refer to docs: https://github.com/TotallyNotChas
 # Float numbers from 1 to 10, upto a single decimal precision
 from decimal import getcontext, Decimal
 # Setting floating point precision to 1 (after decimal point)
+
 getcontext().prec = 2
 amount_list = []
-start = Decimal(1.0)
-for i in range(91):
+start = Decimal(0.1)
+for i in range(100):
     amount_list.append(float(start))
     start += Decimal(0.1)
-
-def get_random_glitch_amount():
-    return choice(amount_list)
 
 def test_loop():
     # A method to stress test
@@ -35,7 +33,7 @@ def test_loop():
         with open('Collections/imglog.txt', 'w') as logtxt:
             while(1):
                 t0 = time()
-                level = get_random_glitch_amount()
+                level = choice(amount_list)
                 """
                  Example of getting a glitched image and saving it,
                  with all default params
@@ -154,7 +152,7 @@ def test_image_to_gif():
     # Now try with increasing the glitch_amount by 1 every time, with cycle set to True
     # glitch_amount will reach glitch_max after (glitch_max - glitch_amount)/glitch_change glitches
     # in this case that's 8
-    # It'll cycle back to 1 after that and keep incrementing by glitch_change again
+    # It'll cycle back to glitch_min after that and keep incrementing by glitch_change again
     glitch_imgs = glitcher.glitch_image('test.{}'.format(fmt), 2, glitch_change=1, cycle=True, gif=True)
     glitch_imgs[0].save('Collections/glitched_test_increment_cycle.gif',
                         format='GIF',
@@ -166,7 +164,7 @@ def test_image_to_gif():
     # Now try with increasing the glitch_amount by -1 every time, with cycle set to True
     # glitch_amount will reach glitch_min after (glitch_min - glitch_amount)/glitch_change glitches
     # in this case that's 1
-    # It'll cycle back to 10 after that and keep incrementing (actually decrementing, in this case)
+    # It'll cycle back to glitch_max after that and keep incrementing (actually decrementing, in this case)
     # by glitch_change again
     glitch_imgs = glitcher.glitch_image('test.{}'.format(fmt), 2, glitch_change=-1, cycle=True, gif=True)
     glitch_imgs[0].save('Collections/glitched_test_decrement_cycle.gif',
@@ -212,9 +210,9 @@ def test_gif_to_gif():
      Example of getting a glitched GIF (from another GIF)
      and saving it
 
-     **`glitch_gif` also returns the duration (of each frame)
+     `glitch_gif` also returns the duration (of each frame)
      in centiseconds and the number of frames in the given GIF, you may
-     use these when saving your GIF too!**
+     use these when saving your GIF too!
 
      We use glitch_level = 2 in this example
      We also are making infinitely looping GIFs
@@ -267,7 +265,7 @@ def test_gif_to_gif():
     # Now try with increasing the glitch_amount by 1 every time, with cycle set to True
     # glitch_amount will reach glitch_max after (glitch_max - glitch_amount)/glitch_change glitches
     # in this case that's 8
-    # It'll cycle back to 1 after that and keep incrementing by glitch_change again
+    # It'll cycle back to glitch_min after that and keep incrementing by glitch_change again
     glitch_imgs, src_duration, src_frames = glitcher.glitch_gif('test.gif', 2, glitch_change=1, cycle=True)
     glitch_imgs[0].save('Collections/glitched_gif_increment_cycle.gif',
                         format='GIF',
@@ -279,7 +277,7 @@ def test_gif_to_gif():
     # Now try with increasing the glitch_amount by -1 every time, with cycle set to True
     # glitch_amount will reach glitch_min after (glitch_min - glitch_amount)/glitch_change glitches
     # in this case that's 1
-    # It'll cycle back to 10 after that and keep incrementing (actually decrementing, in this case)
+    # It'll cycle back to glitch_max after that and keep incrementing (actually decrementing, in this case)
     # by glitch_change again
     glitch_imgs, src_duration, src_frames = glitcher.glitch_gif('test.gif', 2, glitch_change=-1, cycle=True)
     glitch_imgs[0].save('Collections/glitched_gif_decrement_cycle.gif',
@@ -319,7 +317,7 @@ def test_gif_to_gif():
                         save_all=True,
                         duration=DURATION,
                         loop=LOOP)
-
+                        
 if __name__=='__main__':
     # Create the ImageGlitcher object
     glitcher = ImageGlitcher()